BlockMergeTimings and MCMCTimings moved to separate python files
